+++ src/app/events/[eventId]/dates/[dailyEventId]/handy-party/vehicle-auto-assignment/hooks/useCreateCurrentTimeLog.ts
