+++ src/app/events/[eventId]/dates/[dailyEventId]/handy-party/vehicle-auto-assignment/types/vehicleAutoAssignment.type.ts
